app/requirements.txt
asgiref==3.8.1
astroid==3.3.5
dill==0.3.9
Django==4.2.16
django-dotenv==1.4.2
djangorestframework==3.15.2
gunicorn==23.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
psycopg2-binary==2.9.10
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
sqlparse==0.5.2
tomli==2.1.0
tomlkit==0.13.2
typing_extensions==4.12.2
